+++ client/src/App.js
@@ -1,10 +1,11 @@
 import React from "react";
 import { BrowserRouter as Router, Route, Switch } from "react-router-dom";
 import GameLevel from "./components/GameLevel"
+import Paddle from "./components/Paddle"
 
 function App() {
   return (
-    <GameLevel></GameLevel>
+    <Paddle></Paddle>
   );
 }
